NoCoManager/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
PyYAML==6.0.1
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.0
WTForms==3.0.1
cffi==1.15.1
cryptography==41.0.4
Pillow==10.0.1
psycopg2-binary==2.9.9
pycparser==2.21
pyOpenSSL==23.2.0
pypng==0.20220715.0
qrcode==7.4.2
